Conflicts:
	lams_build/3rdParty.userlibraries
	lams_central/src/java/org/lamsfoundation/lams/authoring/template/web/TBLTemplateController.java
	lams_common/src/java/org/lamsfoundation/lams/dbupdates/patch20190119.sql
	lams_common/src/java/org/lamsfoundation/lams/dbupdates/patch20190723.sql
	lams_common/src/java/org/lamsfoundation/lams/dbupdates/patch20190724.sql
	lams_common/src/java/org/lamsfoundation/lams/util/CommonConstants.java
	lams_monitoring/web/tblmonitor/teams.jsp
	lams_tool_assessment/conf/language/lams/ApplicationResources_en_AU.properties
	lams_tool_assessment/src/java/org/lamsfoundation/lams/tool/assessment/dbupdates/patch20190517.sql
	lams_tool_assessment/src/java/org/lamsfoundation/lams/tool/assessment/dbupdates/patch20190809.sql
	lams_tool_assessment/src/java/org/lamsfoundation/lams/tool/assessment/service/AssessmentServiceImpl.java
	lams_tool_lamc/src/java/org/lamsfoundation/lams/tool/mc/model/McUsrAttempt.java
	lams_tool_lamc/web/WEB-INF/tags/ImgButtonWrapper.tag
	lams_tool_lamc/web/WEB-INF/tlds/lams/lams.tld
	lams_tool_lamc/web/authoring/BasicContent.jsp
	lams_tool_lamc/web/learning/CombinedAnswersContent.jsp
	lams_tool_lamc/web/learning/SingleQuestionAnswersContent.jsp
	lams_tool_lamc/web/learning/ViewAnswers.jsp
	lams_tool_lamc/web/monitoring/SummaryContent.jsp
	lams_tool_lamc/web/monitoring/parts/statsPart.jsp
	lams_tool_lamc/web/tblmonitoring/mcqStudentChoices.jsp
	lams_tool_scratchie/src/java/org/lamsfoundation/lams/tool/scratchie/ScratchieConstants.java

+++ lams_tool_laqa/src/java/org/lamsfoundation/lams/tool/qa/web/controller/AuthoringController.java
@@ -107,14 +107,13 @@ public class AuthoringController implements QaAppConstants {
 	ToolAccessMode mode = WebUtil.readToolAccessModeAuthorDefaulted(request);
 	return readDatabaseData(form, request, mode);
     }
-    
+
     /**
      * Set the defineLater flag so that learners cannot use content while we are editing. This flag is released when
      * updateContent is called.
      */
     @RequestMapping(path = "/definelater", method = RequestMethod.POST)
-    public String definelater(@ModelAttribute("authoringForm") QaAuthoringForm form,
-	    HttpServletRequest request) {
+    public String definelater(@ModelAttribute("authoringForm") QaAuthoringForm form, HttpServletRequest request) {
 	Long toolContentID = WebUtil.readLongParam(request, AttributeNames.PARAM_TOOL_CONTENT_ID);
 	qaService.setDefineLater(toolContentID, true);
 
@@ -317,7 +316,7 @@ public class AuthoringController implements QaAppConstants {
 	    @RequestParam String sessionMapID, @RequestParam Long qbQuestionUid) {
 	SessionMap<String, Object> sessionMap = getSessionMap(form, request);
 	SortedSet<QaQueContent> qaQuestions = getQuestions(sessionMap);
-	
+
 	//check whether this QB question is a duplicate
 	for (QaQueContent qaQuestion : qaQuestions) {
 	    if (qbQuestionUid.equals(qaQuestion.getQbQuestion().getUid())) {
@@ -342,7 +341,7 @@ public class AuthoringController implements QaAppConstants {
 
 	return "authoring/itemlist";
     }
-    
+
     /**
      * Shows "This question has already been added" error message in a browser.
      */
@@ -368,6 +367,7 @@ public class AuthoringController implements QaAppConstants {
 	// check whether it is "edit(old Question)" or "add(new Question)"
 	QbQuestion qbQuestion;
 	QaQueContent qaQuestion;
+	boolean isDefaultQuestion = false;
 	final boolean isAddingQuestion = itemIdx == -1;
 	if (isAddingQuestion) { // add
 
@@ -395,8 +395,18 @@ public class AuthoringController implements QaAppConstants {
 	} else {
 	    List<QaQueContent> rList = new ArrayList<>(qaQuestions);
 	    qaQuestion = rList.get(itemIdx);
-	    qbQuestion = qbService.getQuestionByUid(qaQuestion.getQbQuestion().getUid());
+	    long qbQuestionUid = qaQuestion.getQbQuestion().getUid();
+	    qbQuestion = qbService.getQuestionByUid(qbQuestionUid);
 	    qbService.releaseFromCache(qbQuestion);
+
+	    // if it is a default question, do not modify it
+	    // treat it as a new question
+	    isDefaultQuestion = qbService.isQuestionDefaultInTool(qbQuestionUid, QaAppConstants.MY_SIGNATURE);
+	    if (isDefaultQuestion) {
+		String questionContentFolderID = FileUtil.generateUniqueContentFolderID();
+		qbQuestion.setContentFolderId(questionContentFolderID);
+		qbQuestion.setQuestionId(qbService.generateNextQuestionId()); // generate a new question ID right away, so another user won't "take it"
+	    }
 	}
 
 	QbQuestion oldQbQuestion = qbQuestion.clone();
@@ -410,7 +420,8 @@ public class AuthoringController implements QaAppConstants {
 	qbQuestion.setMinWordsLimit(form.getMinWordsLimit());
 	qbQuestion.setFeedback(form.getFeedback());
 
-	int isQbQuestionModified = qbQuestion.isQbQuestionModified(oldQbQuestion);
+	int isQbQuestionModified = isDefaultQuestion ? IQbService.QUESTION_MODIFIED_ID_BUMP
+		: qbQuestion.isQbQuestionModified(oldQbQuestion);
 	QbQuestion updatedQuestion = null;
 	switch (isQbQuestionModified) {
 	    case IQbService.QUESTION_MODIFIED_VERSION_BUMP: {
@@ -439,11 +450,12 @@ public class AuthoringController implements QaAppConstants {
 	//take care about question's collections. add to collection first
 	Long oldCollectionUid = form.getOldCollectionUid();
 	Long newCollectionUid = form.getNewCollectionUid();
-	if (isAddingQuestion || !newCollectionUid.equals(oldCollectionUid)) {
+	if (isAddingQuestion || isDefaultQuestion
+		|| (newCollectionUid != null && !newCollectionUid.equals(oldCollectionUid))) {
 	    qbService.addQuestionToCollection(newCollectionUid, updatedQuestion.getQuestionId(), false);
 	}
 	//remove from the old collection, if needed
-	if (!isAddingQuestion && !newCollectionUid.equals(oldCollectionUid)) {
+	if (!isAddingQuestion && newCollectionUid != null && !newCollectionUid.equals(oldCollectionUid)) {
 	    qbService.removeQuestionFromCollectionByQuestionId(oldCollectionUid, updatedQuestion.getQuestionId(),
 		    false);
 	}
@@ -494,7 +506,7 @@ public class AuthoringController implements QaAppConstants {
 
 	form.setContentFolderID(qbQuestion.getContentFolderId());
 	form.setSessionMapID(sessionMapID);
-	QbUtils.fillFormWithUserCollections(qbService, form, qbQuestion.getUid());
+	QbUtils.fillFormWithUserCollections(qbService, QaAppConstants.MY_SIGNATURE, form, qbQuestion.getUid());
 
 	request.setAttribute(AttributeNames.PARAM_CONTENT_FOLDER_ID, contentFolderID);
 	return "authoring/newQuestionBox";
